+++ frontend/src/components/data-table/chart-transforms/forms/chart-components.tsx
@@ -16,13 +16,13 @@ import {
   SelectItem,
   SelectValue,
 } from "@/components/ui/select";
-import { CHART_TYPE_ICON, COUNT_FIELD } from "./constants";
+import { CHART_TYPE_ICON, COUNT_FIELD } from "../constants";
 import { ErrorBanner } from "@/plugins/impl/common/error-banner";
 import { buttonVariants } from "@/components/ui/button";
 import { type UseFormReturn, useWatch } from "react-hook-form";
 import type { z } from "zod";
-import type { ChartSchema } from "./chart-schemas";
-import { FieldValidators, TypeConverters } from "./chart-spec";
+import type { ChartSchema } from "../chart-schemas";
+import { FieldValidators, TypeConverters } from "../chart-spec";
 import {
   ColumnSelector,
   AggregationSelect,
@@ -32,7 +32,7 @@ import {
   BooleanField,
   type Field,
 } from "./form-components";
-import { CHART_TYPES, type ChartType, SORT_TYPES } from "./types";
+import { CHART_TYPES, type ChartType, SORT_TYPES } from "../types";
 
 export const IconWithText: React.FC<{
   Icon: React.ForwardRefExoticComponent<
